backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.16.0
werkzeug==3.0.1
python-dotenv==1.0.0
Pillow==11.3.0
scikit-learn==1.7.1
joblib==1.5.2
gunicorn==23.0.0
cloudinary==1.44.1
